requirements-optional.txt
datasets
gradio==3.50.2
pandas
requests
albumentations
pymongo[srv]==3.8.0
torch_tb_profiler
scipy